Rewrite Node CLI Auth
