+++ node-cli-auth/package.json
@@ -5,6 +5,6 @@
   "license": "MIT",
   "dependencies": {
     "axios": "^0.21.1",
-    "hwid": "0.2"
+    "node-machine-id": "^1.1.12"
   }
 }
